#3 #4 #5 long идентификаторы, изменил некоторые имена свойств

+++ ServerApps/SmartHomeServer/SmartHome.Data.Interfaces/Models/IBoolSensorHistory.cs
@@ -4,7 +4,7 @@ namespace SmartHome.Data.Interfaces.Models
 {
     public interface IBoolSensorHistory
     {
-        int ID { get; set; }
+        long ID { get; set; }
 
         bool Value { get; set; }
