+++ ServerApps/SmartHomeServer/SmartHome.Data.Interfaces/Models/IEventDeviceHistory.cs
@@ -4,7 +4,7 @@ namespace SmartHome.Data.Interfaces.Models
 {
     public interface IEventDeviceHistory
     {
-        int ID { get; set; }
+        long ID { get; set; }
 
         DateTime CreateDate { get; set; }
